refactor(allocator): extract admin VNC session routes into routes/admin_sessions.py
Moves the three admin troubleshooting routes (peek, connect, release) out
of main.py into their own blueprint, preserving admin_connect_vm's two-level
error-handling (RotationFailed vs. unexpected-exception) exactly. Removes
the now-dead sign_session_cookie_and_redirect import from main.py.
